requirements.txt
audiosegment==0.23.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
decorator==4.4.2
ffmpeg==1.4
Flask==2.0.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
moviepy==1.0.3
numpy==1.20.3
Pillow==8.2.0
pkg-resources==0.0.0
pocketsphinx==0.1.15
proglog==0.1.9
pydub==0.25.1
requests==2.25.1
SpeechRecognition==3.8.1
tqdm==4.60.0
urllib3==1.26.4
webrtcvad==2.0.10
Werkzeug==2.0.0
youtube-dl==2021.4.26
gunicorn==19.9.0
validators-0.18.2
